+++ struts2-actionflow-plugin/src/test/java/com/amashchenko/struts2/actionflow/ActionFlowInterceptorTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2014 Aleksandr Mashchenko.
+ * Copyright 2013-2015 Aleksandr Mashchenko.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -55,6 +55,8 @@ public class ActionFlowInterceptorTest {
                 "HIGHEST_CURRENT_ACTION_INDEX");
         session.put(TestConstants.SKIP_ACTIONS, "SKIP_ACTIONS");
 
+        session.put(TestConstants.OVERRIDE_ACTION_NAME, "OVERRIDE_ACTION_NAME");
+
         // execute clearSession method
         actionFlowInterceptor.clearSession(session);
 
@@ -63,6 +65,7 @@ public class ActionFlowInterceptorTest {
         Assert.assertNull(session
                 .get(TestConstants.HIGHEST_CURRENT_ACTION_INDEX));
         Assert.assertNull(session.get(TestConstants.SKIP_ACTIONS));
+        Assert.assertNull(session.get(TestConstants.OVERRIDE_ACTION_NAME));
     }
 
     /**